requirements.txt
fastapi
uvicorn[standard]
transformers
sentence-transformers
faiss-cpu
torch
python-dotenv
pydantic
pytest
requests
